server/pyproject.toml
[project]
name = "fegg-api"
version = "0.1.0"
description = "FeGG Platform API"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "e2b>=2.9.0",
    "e2b-code-interpreter>=2.4.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.40.0",
]

[tool.uv]
dev-dependencies = []
